#166 started implementing LissiApiService.removeConnection and refactored postForObject (now easier to use by using HttpService.postForObject)
